@@ -5,6 +5,7 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
     "@babel/preset-react": "^7.0.0",
     "@babel/runtime": "^7.21.1",
+    "@rails/webpacker": "^5.4.4",
     "@samvera/ramp": "https://github.com/samvera-labs/ramp.git#f71e579",
     "babel-plugin-macros": "^3.1.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
@@ -19,14 +20,14 @@
     "path-browserify": "^1.0.1",
     "process": "^0.11.10",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.0.0",
-    "react-dom": "^17.0.1",
-    "react-structural-metadata-editor": "https://github.com/avalonmediasystem/react-structural-metadata-editor#avalon-7.7.1",
-    "react_ujs": "^2.4.4",
+    "react": "^19.1.0",
+    "react-bootstrap": "^2.10.9",
+    "react-dom": "^19.1.0",
+    "react-on-rails": "^14.2.1",
+    "react-structural-metadata-editor": "./react-structural-metadata-editor-v2.0.0.tgz",
     "sass": "^1.65.1",
     "sass-loader": "^13.3.2",
-    "shakapacker": "7.0.2",
+    "shakapacker": "8.0.1",
     "stream-browserify": "^3.0.0",
     "style-loader": "^3.3.3",
     "terser-webpack-plugin": "5",
