@@ -109,7 +109,7 @@ class CollectionList extends Component {
 
   handleSubmit = event => {
     event.preventDefault();
-  }
+  };
 
   render() {
     const { filter, sort, filteredResult = [], maxItems, isLoading } = this.state;
@@ -121,7 +121,7 @@ class CollectionList extends Component {
           handleFilterChange={this.handleFilterChange}
           sort={sort}
           handleSortChange={this.handleSortChange}
-	  handleSubmit={this.handleSubmit}
+          handleSubmit={this.handleSubmit}
         />
         {isLoading && <LoadingSpinner isLoading={isLoading} />}
         {(filteredResult.length === 0 && !isLoading) && <CollectionsFilterNoResults />}
